defining the file system repository interface
